Move `Connections` to pools namespace
